adds plotting and loop checks
Check out the isNotSquare method.  Might be useful.

I think the issue we were running into with x%1 is not 0 was an issue of floats being slightly imprecise on tiny numbers.  I'm not sure even the != 0 checks would cover that.  The isNotSquareMethod should cover it.

Also, the fermat factoring algorithms were continuing past n.  The temporary fermfactLoopMaxChecked function illustrates the isNotSquare and loopMax checks working to prevent that.